[BREAKING] Re-implement using libsodium
This makes us independent from the more-ciphers plugin without making
the code more complicated. This also changes the algorithms used as
according to [1] AES-GCM may not the right choice for encrypting
passwords.

With crypto_pwhash() for key derivation and crypto_secretbox_easy() for
account password encryption we use libsodiums abstracted functions now.
They are designed for that specific purpose so they hopefully do the
right thing in terms of encryption out of the box.

To protect the master key as good as possible we use sodium_malloc() and
sodium_free(). This fixes #1. However, due to the implementation of
dialogs and account passwords in libpurple a malicious plugin will still
be able to steel the master password or the account password.

[1]: https://libsodium.gitbook.io/doc/secret-key_cryptography/aead/aes-256-gcm

Signed-off-by: Konrad Gräfe <konradgraefe@aol.com>